+++ Assets/GravityDirectionPack/CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.0.0] - 2022-07-29
+### Changed 
+- Collision detection - try binary search for distance movement correction. 
+  Better wall detection but character stops when should move along wall. Required an improvement.
+- Test case **CharacterControllerTests_Move::IsNotMoveThroughWalls** - 
+  use manual data-rows to reproduce more cases of movement to walls.
+
 ## [v0.0.0] - 2022-07-24
 ### Added
 - Input actions and inputs reader